# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TourOps AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles/me:
    get:
      operationId: getMyProfile
      tags: [profiles]
      summary: Get current user profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMyProfile
      tags: [profiles]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: customerType
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: Deleted

  # ─── SUPPLIERS ───────────────────────────────────────────────────────────
...
  /operations:
    get:
      operationId: listOperations
      tags: [operations]
      summary: List all operations
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of operations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Operation"
    post:
      operationId: createOperation
      tags: [operations]
      summary: Create an operation manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperationInput"
      responses:
        "201":
          description: Created operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
  /operations/{id}:
    get:
      operationId: getOperation
      tags: [operations]
      summary: Get operation with tasks
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Operation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationDetail"
        "404":
          description: Not found
    patch:
      operationId: updateOperation
      tags: [operations]
      summary: Update an operation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperationUpdate"
      responses:
        "200":
          description: Updated operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
  /operations/{id}/tasks:
    get:
      operationId: listOperationTasks
      tags: [operations]
      summary: List tasks for an operation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OperationTask"
    post:
      operationId: createOperationTask
      tags: [operations]
      summary: Create a task for an operation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperationTaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationTask"
  /operations/{id}/tasks/{taskId}:
    patch:
      operationId: updateOperationTask
      tags: [operations]
      summary: Update an operation task
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
        - name: taskId
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperationTaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationTask"
    delete:
      operationId: deleteOperationTask
      tags: [operations]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
        - name: taskId
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: Deleted

  # ─── NOTIFICATIONS ───────────────────────────────────────────────────────
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List all notifications
      parameters:
        - name: unreadOnly
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Profile:
      type: object
      properties:
        id: { type: number }
        clerkUserId: { type: string }
        email: { type: string }
        name: { type: ["string", "null"] }
        role: { type: string, enum: [admin, staff] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, clerkUserId, email, role, createdAt, updatedAt]
    ProfileUpdate:
      type: object
      properties:
        name: { type: string }
        role: { type: string, enum: [admin, staff] }

    # ── Customers ────────────────────────────────────────────────────────
    Customer:
      type: object
      properties:
        id: { type: number }
        name: { type: string }
        company: { type: ["string", "null"] }
        nationality: { type: ["string", "null"] }
        language: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        whatsapp: { type: ["string", "null"] }
        customerType: { type: string }
        travelPreferences: { type: ["string", "null"] }
        dietaryRestrictions: { type: ["string", "null"] }
        accessibilityRequirements: { type: ["string", "null"] }
        passportStatus: { type: string }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, name, customerType, passportStatus, createdAt, updatedAt]
    CustomerInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        company: { type: string }
        nationality: { type: string }
        language: { type: string }
        phone: { type: string }
        email: { type: string }
        whatsapp: { type: string }
        customerType: { type: string }
        travelPreferences: { type: string }
        dietaryRestrictions: { type: string }
        accessibilityRequirements: { type: string }
        passportStatus: { type: string }
        notes: { type: string }
      required: [name, customerType]
    CustomerUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        company: { type: string }
        nationality: { type: string }
        language: { type: string }
        phone: { type: string }
        email: { type: string }
        whatsapp: { type: string }
        customerType: { type: string }
        travelPreferences: { type: string }
        dietaryRestrictions: { type: string }
        accessibilityRequirements: { type: string }
        passportStatus: { type: string }
        notes: { type: string }

    # ── Suppliers ────────────────────────────────────────────────────────
...
    Operation:
      type: object
      properties:
        id: { type: number }
        quotationId: { type: ["number", "null"] }
        tourId: { type: ["number", "null"] }
        customerId: { type: number }
        startDate: { type: ["string", "null"] }
        status: { type: string }
        assignedTo: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        completionRate: { type: number }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, customerId, status, completionRate, createdAt, updatedAt]
    OperationDetail:
      type: object
      properties:
        id: { type: number }
        quotationId: { type: ["number", "null"] }
        tourId: { type: ["number", "null"] }
        customerId: { type: number }
        startDate: { type: ["string", "null"] }
        status: { type: string }
        assignedTo: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        completionRate: { type: number }
        tasks: { type: array, items: { $ref: "#/components/schemas/OperationTask" } }
        customer: { $ref: "#/components/schemas/Customer" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, customerId, status, completionRate, tasks, customer, createdAt, updatedAt]
    OperationInput:
      type: object
      properties:
        quotationId: { type: number }
        tourId: { type: number }
        customerId: { type: number }
        startDate: { type: string }
        status: { type: string }
        assignedTo: { type: string }
        notes: { type: string }
      required: [customerId]
    OperationUpdate:
      type: object
      properties:
        startDate: { type: string }
        status: { type: string }
        assignedTo: { type: string }
        notes: { type: string }
    OperationTask:
      type: object
      properties:
        id: { type: number }
        operationId: { type: number }
        title: { type: string }
        status: { type: string }
        priority: { type: string }
        dueDate: { type: ["string", "null"] }
        assignedTo: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        completedAt: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required: [id, operationId, title, status, priority, createdAt]
    OperationTaskInput:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        status: { type: string }
        priority: { type: string }
        dueDate: { type: string }
        assignedTo: { type: string }
        notes: { type: string }
      required: [title]
    OperationTaskUpdate:
      type: object
      properties:
        title: { type: string }
        status: { type: string }
        priority: { type: string }
        dueDate: { type: string }
        assignedTo: { type: string }
        notes: { type: string }
        completedAt: { type: string }

    # ── Notifications ────────────────────────────────────────────────────
    Notification:
      type: object
      properties:
        id: { type: number }
        type: { type: string }
        title: { type: string }
        message: { type: string }
        isRead: { type: boolean }
        severity: { type: string }
        relatedId: { type: ["number", "null"] }
        relatedType: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required: [id, type, title, message, isRead, severity, createdAt]

    # ── Exchange Rates ───────────────────────────────────────────────────
...